review: fit-solve double-precision b term + doc/test-convention fixes
1. src/AtlasPacking.cpp line 832: restore pre-branch double-precision math
   for b accumulation — cast crects[c].w to double before summing.
2. docs/BENCHMARKS.md line 290: fix non-existent knob name
   developable_repair_carve_rings → repair_carve_rings.
3. CHANGELOG.md line 37-38: reword triangle-coverage claim; atlasbench
   does not wire AtlasResult::coverage, only its own occupancyTri metric.
4. tests/FlattenTest.cpp line 913: convert hard ASSERT to GTEST_SKIP()
   pattern, matching sibling mesh.ply tests.
5. src/AtlasCharting.cpp line 481: document PairKey conservatism trade-off
   in min-fid veto over-approximation.

+++ src/AtlasPacking.cpp
@@ -829,7 +829,7 @@ AtlasResult PackAtlas(Mesh& mesh,
 				continue; // degenerate charts stay a fixed 1-texel slot, unscaled
 			const float padC = ChartPad(params, crects[c].w, crects[c].h, chartFaces[c], pad);
 			a += static_cast<double>(crects[c].w) * crects[c].h;
-			b += 2.0 * static_cast<double>(padC) * (crects[c].w + crects[c].h);
+			b += 2.0 * static_cast<double>(padC) * (static_cast<double>(crects[c].w) + crects[c].h);
 			cc += 4.0 * static_cast<double>(padC) * padC;
 		}
 		const double R = static_cast<double>(params.resolution);
@@ -1017,8 +1017,7 @@ AtlasResult PackAtlas(Mesh& mesh,
 		const TexCoord& t0 = mesh.faceTexcoords[fi * 3 + 0];
 		const TexCoord& t1 = mesh.faceTexcoords[fi * 3 + 1];
 		const TexCoord& t2 = mesh.faceTexcoords[fi * 3 + 2];
-		triArea += 0.5 * std::abs(static_cast<double>(t1.x() - t0.x()) * (t2.y() - t0.y())
-		                          - static_cast<double>(t2.x() - t0.x()) * (t1.y() - t0.y()));
+		triArea += 0.5 * std::abs(static_cast<double>(t1.x() - t0.x()) * (t2.y() - t0.y()) - static_cast<double>(t2.x() - t0.x()) * (t1.y() - t0.y()));
 	}
 	result.coverage = (numPages > 0)
 	                      ? std::min(1.f, static_cast<float>(triArea / numPages))
